requirements.txt
connexion>=2.2.0
connexion[swagger-ui]
python-jose==3.0.1
python-jose[cryptography]
six>=1.9
Flask>=0.10.1
pymongo>=3.7.2
swagger-ui-bundle>=0.0.3
